@@ -1,51 +1,58 @@
 {
+  "author": "Rodrigo Dalmas",
   "name": "react-hot-webpack-express-boilerplate",
   "version": "1.0.0",
   "description": "Boilerplate for ReactJS project with hot code reloading and express",
+  "license": "MIT",
   "scripts": {
-    "start": "node ./server/server.js",
+    "start": "nodemon --exec babel-node ./server/server.js",
     "lint": "eslint src",
     "prebuild": "rimraf dist",
+    "build:dev": "cross-env NODE_ENV=development webpack -p --config ./config/webpack.config.js",
     "build": "cross-env NODE_ENV=production webpack -p --config ./config/webpack.config.production.js",
     "postbuild": "copyfiles index.html dist"
   },
-  "keywords": [
-    "react",
-    "reactjs",
-    "boilerplate",
-    "hot",
-    "reload",
-    "hmr",
-    "live",
-    "edit",
-    "webpack"
-  ],
-  "author": "Brandon Jones <brandon.jones1985@gmail.com> (http://github.com/bjonesy)",
-  "license": "MIT",
-  "devDependencies": {
-    "babel-core": "6.25.0",
-    "babel-loader": "7.1.1",
-    "babel-preset-es2015": "6.24.1",
-    "babel-preset-react": "6.24.1",
-    "babel-preset-stage-2": "6.24.1",
-    "copyfiles": "^1.2.0",
-    "cross-env": "5.0.1",
-    "react-hot-loader": "3.0.0-beta.7",
-    "rimraf": "^2.6.0",
-    "webpack": "3.3.0",
-    "webpack-dev-server": "2.5.1"
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/rdalmas/react-hot-webpack-express-boilerplate.git"
   },
+  "presets": [
+    "es2015"
+  ],
   "dependencies": {
+    "axios": "^0.19.2",
     "babel-eslint": "7.2.3",
-    "css-loader": "^0.28.0",
+    "css-loader": "^3.5.2",
     "eslint": "4.2.0",
     "eslint-plugin-react": "7.1.0",
+    "express": "^4.17.1",
     "node-sass": "^4.5.2",
-    "react": "15.6.1",
-    "react-dom": "15.6.1",
+    "qs": "^6.9.3",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "sass-loader": "^6.0.3",
-    "style-loader": "^0.16.1",
-    "webpack-dev-middleware": "1.11.0",
-    "webpack-hot-middleware": "2.18.2"
+    "style-loader": "^1.1.3"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/node": "^7.8.7",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.9.0",
+    "@babel/preset-env": "^7.9.5",
+    "@babel/preset-react": "^7.9.4",
+    "@babel/register": "^7.9.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-loader": "7.1.1",
+    "copyfiles": "^1.2.0",
+    "cross-env": "5.0.1",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "nodemon": "^2.0.3",
+    "react-hot-loader": "4.12.20",
+    "rimraf": "^2.7.1",
+    "webpack": "4.29.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-middleware": "^3.7.2",
+    "webpack-hot-middleware": "^2.25.0"
   }
 }
