@@ -2,19 +2,19 @@ import './assets/scss/base.scss';
 import { AppContainer } from 'react-hot-loader';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './containers/App';
 
-const rootEl = document.getElementById('root');
+import App from "./pages/App.jsx";
+
 const render = Component =>
   ReactDOM.render(
     <AppContainer>
       <Component />
     </AppContainer>,
-    rootEl
+    document.getElementById('root')
   );
 
 render(App);
-if (module.hot) module.hot.accept('./containers/App', () => {
+if (module.hot) module.hot.accept('./pages/App.jsx', () => {
   try {
     render(App)
   } catch (e) {
